renamed elaradb module to elara
