+++ elara/shared.py
