+++ elara/strings.py
